Add API to get and set server config path
This introduced a new DBus call on Server to get path to its
config path. This is useful when one wants to figure out the
config path of a Server running via testrunner, i.e. having
XDG_CONFIG_DIRS randomly set in /tmp.

Secondly it introduced an API in XdgBaseDirs which allows
overriding the config dir path - this is useful when you
want to connect to a server which is running with different
XDG_CONFIG_DIRS than yours, e.g. in AkonadiConsole.