AgentType: fix X-Akonadi-Custom- value parsing
Tread the default value as QString, not QStringList. The string value
is stored in QVariant and can then be converted from boolean, number
via QVariant::to*() methods. This does not work if the value in QVariant
is stored as QStringList.

This fixes the isScopeLocal() hack for Indexing Agent in FetchHelper.