Make COPY and MOVE handlers more interactive
When copying or moving large batch of items they all have to
go through ItemRetriever first to make sure all payload parts
are locally cached. This can take a lot of time, so to make
the operation seem more responsive, we don't wait for
ItemRetriever to finish retrieval for all the items, but
processes every batch as soon as the ItemRetriever reports
it as finished.