MOVE: fix concurrency bug and improve performance on complex moves
The MOVE handler would emit the notification before storing the updated
data in the database. Sometimes this  has lead to the resource getting
the notification and fetching the Item before the Server would update it
in the database. Specifically in case of IMAP resource this would hit a
code path that checks whether the moved Item has a correct parent
set and discard the change if not.

This patch makes the changes to be stored in DB right away and the
notification is emitted after that. It also does only a single iteration
over the whole list of Items to be moved, which should be faster in
case of batched moves. Finally it uses just a single SQL query to
reset RID of all items that were moved between resources instead
of udpating each Item individually.