Display selected, newly created tags
In some cases, it is necessary to set newly created tags as selected for the
TagView. The text field in TagView only displays tags that are part of its
TagModel. This change allows TagWidget to update its view when selected tags
that did not exist initially were added to the TagModel.

Differential Revision: https://phabricator.kde.org/D2433