Fix MTIME handling accross multiple backends
MySQL and PSQL don't support milliseconds in datetime columns, so
we need to strip milliseconds in STORE response before sending it
back, otherwise the MTIME send by STORE does not match MTIME of a
separately fetched from the database.

Secondly, MySQL and SQLite QtSql drivers assume that the datetime
stored in DB is local time (even though we stored as UTC), while
PSQL driver assumes it's UTC and converts to local time, so we
need to handle this separetly for each backend.