ErrorOverlay: fix dialogs being disabled by the mainwindow
Testcase: open "jump to folder", cancel, restart akonadi,
reopen the dialog, it was disabled.

This happened because disabling the window actually disables
the child dialog too, so the saved state for the dialog was
"disabled". The fix is to save the "explicitly disabled" state
rather than isDisabled(). Thanks to Volker for helping me figure
this out.

Differential Revision: https://phabricator.kde.org/D2717