[BIC] Let ETM be constructed with Monitor instead of ChangeRecorder
There is zero reason to require ETM to be constructed with
ChangeRecorder, when none of the recording features are actually
used there. Instead let ETM be created with Monitor.

Monitor is a baseclass of ChangeRecorder so no existing code needs to
be changed, but it's a binary incompatible change, so everything
linking Akonadi must be rebuilt.