Introduce concept of database generation
Generation is an integer which is guaranteed to never change as long as
the database backend is not removed and re-created. If that happens it
is guaranteed that the new generation identifier will be higher than
the previous one.

Client applications can access it via ServerManager::generation(). The
purpose of the generation number is to make it possible for applications
to detect when the database was recreated and some of their configuration
must be invalidated (e.g. collection IDs stored in config files).