ChangeRecorder: fix a bug in Relation ntf serialization
The serialization code had a typo that caused it to serialize a qint64
(8 bytes) instead of empty QByteArray (4 bytes), so the deserialization
code got stuck on reading completely broken values. This was causing the
CR to get stuck in an endless loop (prior to fix from David) or reading
bogus data.