Reset Item RID on both inter- and intra-resource MOVE
This should significantly help to mitigate the "Multiple Merge
Candidates" error during sync. One of the major reasons why MMC
happens is when user moves an Item belonging to an IMAP resource
to another Collection. On IMAP RID corresponds to IMAP UID. If
you move such an Item into another Collection which has IMAP
UIDNEXT greater than the source Collection, then there's a high
chance for RID conflict until the Resource replays the change
and updates the RID of the moved Item.

This change reset RID of all moves Items and only emits the RID
as part of the change notification. The Monitor then retrieves the
Item without the RID and restores the RID from the Notification,
so this is absolutely transparent to Resources.

BUG: 338658
FIXED-IN: 5.4.1