Monitor: fix handling of CollectionRemoved notifications
Many resources expect to always get a full ancestor chain of
changed Collections. CollectionRemoved notification however has
a "fast path" which by-passes retrieval of the removed Collection
from Akonadi, because the Collection no longer exists there. This
howerver means that the Collection is only reconstructed from data
provided in the change notification and thus does not have the
ancestor chain. This change removes the "fast path" from Monitor
but instead of retrieving the removed Collection it retrieves its
parent (and its ancestor chain) so the resource implementation
gets a full ancestor chain.