Fix database deadlock introduced in commit d54d1f241
The early-return optimization in SearchManager::searchUpdateResultsAvailable()
did not commit or rollback the transaction that was opened in that method.
If the transaction acquired an exclusive lock it would block all other DB
connections from acquiring exclusive transactions on that table forever.
This only affects SQLite, MySQL/PSQL seems to be able to deal with it
thanks to row locking.