Only update Item's atime if it's relevant
If the parent Collection does not have expiration timeout, then
the atime value will never be used, so there's no point wasting
time on updating it. This check however only happens when FETCH-
ing entire Collection, not if we are fetching on per-item basis,
because then the check would be way too expensive.