DataStream: throw exception if device is null
During Connection destruction it can happen that the socket is
deleted, but the actual handler is still running and tries to
write to that socket (in which case it passes a nullptr to
DataStream). This is caused by the issue described in #376822.

To prevent the crash in this case (until the root cause is fixed)
the DataStream throws an exception if its used without a valid
device. This will effectively cause execution to leave the Handler
and return.

BUG: 376385
FIXED-IN: 5.4.3