Introduce Foreign payloads
In addition to having internal and external payload parts, we now
have also foreign payload parts. While external payload parts are
stored outside of the database, they are still managed by Akonadi.
Foreign payloads are also stored outside of the database, but they
are managed by the resource. In other words foreign payload part
points to the file stored in the backend storage, like maildir or
icaldir.

This hugely speeds up handling of maildir emails whose payload is
expired most of the time and opening those emails requires Akonadi
to request the resource to upload the payload just to expire it
in couple minutes again causing a lot of IO and delays. With foreign
payloads the client can read the payload part directly from the
backend storage.