ResourceBase: cancelTask() when PrepareItemsRetrieval job returns 0 items
It can happen that an Item is removed from the Server between the Server
sending a requestItemDelivery() request and the Resource scheduling the
retrieval. In such case the ItemFetchJob returns 0 Items. Instead of
passing an empty list into the resource implementation, just cancel the
task.

Also fixes a crash in Debug builds due to Q_ASSERT unconditionally
calling items.first().