Fix trx deadlock when updating Search collection
Only affects SQLite: when we are in a transaction while modifying
a collection, calling SearchManager::updateSearch() can trigger a
transaction deadlock, because the search result delivery mechanism
will try to also get an exclusive lock in order to write the new
results into the database.

To workaround that we only request the search update AFTER we commit
all changes to the Collection itself.