don't CreateItem when it is already in the DB. Use ModifyItem instead.
To avoid running into the "multiple merge candiate" problem when
akonadi searches items by RID and for whatever reason there are
really duplicate RIDs, instead when we already have an item ID from
the DB, avoid using a Create job at all and answer a fetch with
a Modify Job. Akonadi will then query this item by the given id,
which is unique and we can never get multiple items in return.
Also a query by id is faster since it's indexed.

Differential Revision: https://phabricator.kde.org/D5652