Send Collection as part of CollectionChangeNotification
The Monitor can still decide to retrieve the Collection from Akonadi
in certain cases, but generally in most cases the Monitor should emit
the notification with Collection created directly from the payload of
the Notification.

We still need to keep around the EntityCache and Collection retrieval
code in Monitor so that we can retrieve Item parent Collections and
missing parts for Collections reconstructed from previous version of
ChangeRecorder journal.