Add aggregated collection fetch scope into NotificationManager
The aggregated fetch scope ensures that the Collection has all the
data that were requested by at least one subscriber. This happens
before the Notification is dispatched to individual subscribers, so
we only retrieve the data once. It is possible that some subscribers
will receive more data than they requested, but that's perfectly
OK.