Workaround a weird misbehaviour of SQLite when checking tag RID
For some reason "WHERE remoteId = 'rid-with-hyphen'" would not match
any rows even if such row exists in TagRemoteIdResourceRelation. This
only seems to affect a specific case in Zanshin's akonadistoragetest.
I couldn't reproduce this on a clean SQLite database nor could  I
reproduce it with manually created entries in the test's instance DB,
but I could confirm that with data created by the test itself.

For now the problem is worked-around by using "LIKE" instead of "=",
which matches the strings correctly.