Fix "Session destroyed while still in pool" assertions
When the system is not using NetworkManager, the code should fallback
to the old QNetworkConfigurationManager code. But the previous patch
also asked NetworkManager for connectivity even when
QNetworkConfigurationManager is used but the system is online. This
then resulted in setOnlineInternal never setting the state to online.
All resources are offline and some even assert and crash.