LIST: correctly return mimetypes for all Collections
LIST was only returning mimetypes for Collections that were resolved in the
initial filter pass. All Collections that were resolved in the second pass
to complete the tree were missing mimetypes, because their IDs were not
passed to the mimetype query. This in included mainly Collections that
don't match the initial mimetype filter, like top-level Collections.

This caused, among other things, KMail to refuse to create Collections
under the top-level Collection, because the top-level Collections did
not match KMail's mimetype filter (inode/directory).

BUG: 350219
FIXED-IN: 5.6.2