Initial implementation of server-based Indexer
Indexer lives in a separate thread and has a queue of tasks to process
that is persistently stored accross restarts, so that no indexing
task is lost in case of a crash or restart.

The Indexer passes the data directly to an Indexer plugin, provided
generally by akonadi-search, which then writes the data into its
Xapian Store.