EntityOrderProxyModel: adapt base class for reordering favorites
Summary:
In order to use an EOPM on top of the favorites model, we need
the config key to be fixed (e.g. "0") rather than "the parent
collection" in ETM, since the favorites are a flat list of individually
picked collection, with a different parent in ETM. We want a single
order. So the new virtual parentCollection is reimplemented in a new
proxy derived class in mailcommon, for ordering favorites.

Test Plan:
With the mailcommon and kmail patches on top of this,
I can reorder favorites using drag-n-drop and the order is persisted.

Reviewers: dvratil, mlaurent

Reviewed By: mlaurent

Subscribers: #kde_pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D8883