AKAPPEND: Merge by RID if matching Item with empty GID exists during GID merge
This is more or less a workaround for a real-world issue that we've
been seeing for a while now. Sometimes an Item is initially created
with an empty GID, but upon next sync the same Item with the same
RID has a non-empty GID. During GID merge this creates a RID duplicate,
since the GIDs did not match.

To workaround that problem, the AKAPPEND merging code falls back to
RID merge if it finds that an Item with matching RID but empty GID
exists during GID merge.