Don't consider ItemFetchScope with only ancestor retrieval set to be empty
The default IFS in Monitor is empty (default-constructed). If user only
changes ancestor retrieval depth, then the IFS must no longer be
considered empty - an empty IFS triggers a special code path in the
Monitor that bypasses EntityCache, and so we won't get the full
Item data (like flags) in the ItemAdded or ItemMoved notifications.