EntityOrderProxyModel: support adding a favorite at a specific position
Summary:
If EntityTreeModel::modelIndexForCollection() doesn't find the
collection we're dropping, then insert it first (via the base class
dropMimeData) and then look for it again, to handle it as a move.

I hope this doesn't affect other EOPM users than the favorites model;
but when reordering a full ETM surely the dropped folder is somewhere in the
ETM.

Test Plan:
When using the ListView mode for the favorites, this allows to
directly drop a new folder (from the folder tree) to a specific position,
as shown by the "between items" drop indicator.

Reviewers: dvratil, mlaurent

Reviewed By: dvratil

Subscribers: #kde_pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D9417