Don't try to move log file which might be in use
Summary:
While this workaround fixes the crash, the root of the problem is that every process liking that code tries to rename the file.

In the most basic case this is akonadi_control and akonadiserver.

On Windows it gets even more complicated as we might have multiple instances.

Reviewers: dvratil

Reviewed By: dvratil

Subscribers: #windows, #kde_pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D9796