Port from XdgBaseDirs::findExecutableFile() to QStandardPaths::findExecutable()
The added functionality of XdgBaseDirs to look for agent binaries in
alternative locations on Windows is no longer necessary as the new way
of deploying applications on Windows is in a single bundle so the
binaries will be always in the same installation prefix.

All the other functionality is already provided by QStandardPaths.