TransactionSequence: fix rollback not finishing properly.
Summary:
emitResult was missing in the case where we are done cleaning up the
subjobs.

Test Plan:
F5 in a large folder in kmail, then pressing the Abort button
in the middle of the sync. It used to just block the sync, while now it's
possible to resume afterwards.

Reviewers: dvratil

Subscribers: #kde_pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D12426