ActionStateManager: separate selected collections from favorites.
Summary:
When filtering the folder tree (e.g. type "zzz"), a favorite can
be selected, but without any selection in the folder tree.
So the code should stop assuming that both selections are the same.

That testcase is still broken when it comes to actually reading
that favorite folder, but at least the popupmenu actions work now.

Test Plan: As described above, in kmail.

Reviewers: mlaurent, dvratil, vkrause

Reviewed By: vkrause

Subscribers: kde-pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D13893