Enrich pending notifications from ItemModifyJob with RID
This is a bit of a workaround to handle one specific case when we have two
subsequent notifications to a Resource: ADD and MODIFY/MODIFYFLAGS/MOVE etc.
The ADD notification will cause the resource to assign RemoteID to the Item
and issue an ItemModifyJob. The second pending notificiation, however, gets
discarded by the Resource because it's missing RID (it wasn't known by the
time the notification was sent) and is not an ADD notification.

To solve this, we hook into the ItemModifyJob, which the Resource uses to
upload a new RID to the Server and we notify all Monitors through the
ChangeMediator to check if they have any pending notifications for the
particular Item and to add the newly uploaded RID to them, so that the
Resource can process them.