Notifications: retrieve Item when RID is missing
When Akonadi server does not have a RID when it's trying to emit
a notification, it sets a flag instead and sends a notification with
only the Item ID. The client's Monitor will then fallback to the
pre-notification-payload method of retrieving the Item on its own,
ensuring that the updated version with RemoteID is retrieved as well.

This fixes a bug when if a client created a new Item and immediately
modified it, the Modify notification did not have a RID, because the
Resource haven't processed the Add notification yet and haven't
uploaded a RID back to the Akonadi. The Resource then wasn't able to
act on the Modify notification.