Monitor: use (dis)connectNotify to track listeners
Use QObject::(dis)connectNotify() events to track which of Monitor's
signals have someone connected to them. This should be more efficient
than using QObject::receivers() since it only costs us a single hash
table lookup and we can use pointer-to-member-function to refer to the
signals instead of the old-style SIGNAL() syntax.