Server: create new Collection in a single step
We can create a new Collection and append its mimetypes and attributes into
the DB in a single step, so that the entire Collection is present in the
Add change notification.
This fixes a bug when new Collections wouldn't appear sometimes because they
did not have any mimetype set in the notification and thus did not pass an
ETM (or some other) mimetype filters.