Add support for Foreign Keys to SQLite
SQLite supports Foreign Keys for a while now, so we can make use of them to
enforce consistency on SQLite as well.

Note that SQLite does not support adding foreign keys onto existing tables,
so they all have to be recreated (data are preserved), which may take a
huge amount of time.