Monitor: rebuild Item from ntf if it's mustRetrieve but IFS is empty
When Monitor's IFS is empty and it receives an Item notification
with mustRetrieve set to true, then the retrieval is bypassed, but
the later on the code attempts to extract a result from cache which
results crash as we try to access elements in an empty list.

The fix is to send as much metadata as we can in the mustRetrieve
notification, and the Monitor needs to rebuild the Item from the
metadata instead of attempting retrieval.

BUG: 397131
FIXED-IN: 5.9.0