ItemSync can be canceled but not defered. Add assert and debug output.
Summary:
And turn the warning in rollback to a debug, this isn't a problem when
it happens because the task is canceled (e.g. interval-sync with no network
available).

Test Plan:
I've been running with this for some time and I never hit
the assert (after fixing the imap resource and the FatCRM resources to do this
correctly). Let's hope no other resource hits this. But OTOH, if they were deferring
an itemsync, they would get stuck forever, the job never finishing...

Reviewers: dvratil, vkrause

Reviewed By: dvratil

Subscribers: kde-pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D14986