Clean up TagFetchScope handling on client and server side
The fetch scope on the client side was using some hacks to list the
parts (GID, RID, ...) that should be retrieved and on the Item
retrieval part this was coupled with ItemFetchScope. This change
adds standalone TagFetchScope to FetchItemsCommand as well as
ItemFetchJob and similar APIs, allowing users to fully customize
the retrieval.