Fix regression in tag fetching for notifications
Commit 9ea06c01ba introduced some clean-up in tag
fetching. Unfortunately with it came a regression which caused tags
not to be fetched as part of notifications - no matter how the fetch
scope is configured.

The problem lies in the fact that the FetchHelper class, which
populates items based on the configured fetch scope still uses the
Tags flag in the item fetch scope to determine if tags should be
fetched at all and only if this flag is set the tag fetch scope is
examined to determine what tag information to include.

On the notification side however support for handling this flag in the
AggregatedItemFetchScope was removed, which causes the flag never to
be set for the item fetch scope, which in turn means that no tags will
ever be fetched for notifications regardless of tag fetch scope
setting.

This commit brings back handling of the Tags flag in the item fetch
scope thereby partially reverting commit 9ea06c01ba.

Reviewers: #kde_pim

Reviewed By: dvratil

Differential Revision: https://phabricator.kde.org/D16512