Rename Handlers to match the entity, operation and command names
Historically, the protocol command handlers were called after the IMAP
commands from old ASAP. Most of the names do not map to names of
commands in the new protocol, some of them are even quite confusing
(like Remove and Delete handlers). This change renames all handlers
to match their purpose and map more nicely to names of the protocol
commands as well as the client-side jobs.