Handle query prepare errors in QueryBuilder::exec
Summary:
It is way better to handle these errors here as early as possible. It
means seeing error descriptions which actually make sense (in e.g. SQLite
case, when SQLite limits or syntax errors are hit) and probably it should
help to avoid nasty crashes later on in exec() phase.

Reviewers: #kde_pim, dvratil

Reviewed By: #kde_pim, dvratil

Subscribers: dvratil, kde-pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D17877