actually look for mysql binaries in PATH
Summary:
the new helper findExectuable first tries to find the executables in the
default location. which is all the sane distro/system paths and will 99%
yield a result.
should it come back negative we'll fall back to the hardcoded list of
paths as previously used.

this fixes binary lookup when the entire dir tree is located *not* under /
(as with snap and I think also appimage)

Reviewers: dvratil

Reviewed By: dvratil

Subscribers: vandenoever, kde-pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D18456