autotests: remove QEXPECT_FAIL, the mysql bug got fixed.
Summary:
It was resulting in XPASS for me and on CI, so this will actually fix CI.
I'm too lazy to add a mysql version check, not worth it in the long run.
Leaving comments in case anyone hits this with an old mysql.

Not sure what the impact of the fix has on the rest of kdepim though.
Maybe it allows removing workarounds?

Test Plan: ctest -R autoincrement

Reviewers: dvratil

Subscribers: kde-pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D18557