Properly clean up scopes when deregistering a subscriber.
Summary:
It was done "by hand", for only a few setters; use apply() to do this
consistently for all of them.

New unittest notificationmanagertest
This requires adding a special mode for unittests in AkThread, to be
able to call methods in NotificationManager from the main thread.

Test Plan: bin/notificationmanagertest

Reviewers: dvratil

Subscribers: kde-pim

Differential Revision: https://phabricator.kde.org/D18672