AggregatedFetchScope: fix fetchAllAttributes for tags.
Summary:
This is a third case for aggregation: the default is *true*,
so any subscriber with default settings was NOT calling
AggregatedTagFetchScope::setFetchAllAttributes, so
d->fetchAllAttributes > 0 was false, unexpectedly.

Count the number of subscribers who call with false, and return
false only if all of them agree with that.

Test Plan: tagtest

Reviewers: dvratil

Subscribers: kde-pim

Differential Revision: https://phabricator.kde.org/D18695