Fix tag change notification shipping a bogus name.
Summary:
The code doing a SQL query was preparing a condition, but wasn't
actually using it. So this was listing all attributes from
ALL tags (instead of the desired one) and the last one would win.

Bug initially detected by zanshin's unittests.

Test Plan:
Added a new unittest in tagtest which failed before this fix
(the notification contained the name of the second tag)

Reviewers: dvratil

Subscribers: ervin, kde-pim

Differential Revision: https://phabricator.kde.org/D18709