Fix intermittent race in collectionattributetest
Summary:
Any CollectionModifyJob of a collection with attributes, would send the
full list of attributes, even if they hadn't changed. This leads to the
following problem:
* collectionattributetest creates a collection with an attribute
* it then proceeds to modify the attribute
* then the knut resource's collectionAdded() wants to add a
remoteid to that collection, so it creates a ModifyJob of its own
(and that's where the bug happens: it includes all attributes, with the
orig value).

Therefore the old value overwrites the new value again.

Fixed by only sending attributes if they have changed.

Test Plan: collectionattributetest in a loop, doesn't fail anymore

Reviewers: dvratil, vkrause

Reviewed By: dvratil

Subscribers: kde-pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D19290